+++ changelog.md
@@ -3,9 +3,10 @@
 ## 2.4.0
 
 - Fixed conflict handling for debug keys on 1.21.11.
+- Added a new activation mode for key release.
 - Added a command to switch profiles.
 - Added an optional list entry name field for macros.
-- Changed default Conflict Strategy from SUBMIT to ASSERT.
+- Changed default Conflict Strategy from Submit to Assert.
 - Increased command/message field length limit from 512 to 4096.
 
 ## 2.3.17