+++ changelog.md
@@ -1,9 +1,9 @@
 # Changelog
 
-## 2.4.0
+## 2.4.0-beta.1
 
 - Fixed conflict handling for debug keys on 1.21.11.
-- Added a new activation mode for key release.
+- Added a new activation mode for key Release.
 - Added a command to switch profiles.
 - Added an optional list entry name field for macros.
 - Changed default Conflict Strategy from Submit to Assert.